+++ vector/src/gplay/AndroidManifest.xml
@@ -4,15 +4,7 @@
 
     <application>
 
-        <receiver android:exported="true"  android:enabled="true"  android:name=".gplay.push.fcm.VectorMessagingReceiver">
-            <intent-filter>
-                <action android:name="org.unifiedpush.android.connector.MESSAGE"/>
-                <action android:name="org.unifiedpush.android.connector.UNREGISTERED"/>
-                <action android:name="org.unifiedpush.android.connector.NEW_ENDPOINT"/>
-                <action android:name="org.unifiedpush.android.connector.REGISTRATION_FAILED"/>
-                <action android:name="org.unifiedpush.android.connector.REGISTRATION_REFUSED"/>
-            </intent-filter>
-        </receiver>
+
 
     </application>
