+++ vector/src/main/java/im/vector/app/core/pushers/UPHelper.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package im.vector.app.push.fcm
+package im.vector.app.core.pushers
 
 import android.app.Activity
 import android.content.Context
